Move distribute to own file
